# erc8004-agents/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ERC-8004 Agent Registry API",
    "description": "On-chain interaction patterns for ERC-8004 trustless AI agent identity. This describes the contract interaction surface for the IdentityRegistry, ReputationRegistry, and ValidationRegistry deployed on BSC, Ethereum, and EVM-compatible networks. While the web UI is a static site, agents interact via direct JSON-RPC calls to the smart contracts.",
    "version": "1.0.0",
    "contact": {
      "name": "ERC-8004",
      "url": "https://github.com/nirholas/erc8004-agent-creator"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-logo": {
      "url": "https://erc8004.agency/favicon.svg"
    }
  },
  "servers": [
    {
      "url": "https://data-seed-prebsc-1-s1.bnbchain.org:8545",
      "description": "BSC Testnet JSON-RPC"
    },
    {
      "url": "https://bsc-dataseed.bnbchain.org",
      "description": "BSC Mainnet JSON-RPC"
    },
    {
      "url": "https://eth.llamarpc.com",
      "description": "Ethereum Mainnet JSON-RPC"
    },
    {
      "url": "https://rpc.sepolia.org",
      "description": "Ethereum Sepolia JSON-RPC"
    }
  ],
  "paths": {
...
    "/agent/{agentId}": {
      "get": {
        "operationId": "getAgent",
        "summary": "Look up an agent by token ID",
        "description": "Reads tokenURI, ownerOf, and metadata from the IdentityRegistry for a given agentId.",
        "tags": ["Identity"],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Agent token ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agentId": { "type": "string" },
                    "owner": { "type": "string" },
                    "uri": { "type": "string" },
                    "metadata": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}